ball V0.4
Tested ball in main, added an update method that draws the ball and updates the state.

Noticed something: Planet acceleration goes to infinite values over time, 
I think this is because the ball approaches the center of the planet and distance approaches 0. 
Might be fixed when collisions are implemented to stop the ball from going inside?

Also should probably limit speed/acceleration/position so the ball doesn't jump too much. 
Could be interesting to allow the ball to go offscreen or maybe that's a reset?